Keeping deck in memory until change in Deck table in Realm DB
Deck table is change in: updateDeck, addDeck, removeDeck x2 and removeDecksFromDatabase. In this methods deckDBChanged is set to true, so decks will be loaded from database to memory. In loadDecksFromDatabase deckDBChanged is set to false, couse deck array is equal to database Deck table. With next decks() call, decks won't be loaded.